Add support to Gen 3+ to randomize held items for trainer's pokes
This adds two list of items per Gen, one for all [1] consumable held items and one for all [1] held items. This adds the options to add items from either of these lists to boss trainers, important trainers, and/or regular trainers. There is a disabled UI element for future work on "sensible" items, which would use a smaller pool of items (e.g. include the Yache berry in the pool only if the Pokemon is weak to ice).

There had been a smaller list of items used for the Totem pokemon in Gen 7; this has been combined with this new code.

[1] We exclude held items that affect only things for the player (e.g. amulet coin, macho brace ), affect only one species of Pokemon (e.g. light ball, lucky punch), or affect only one move (e.g. douse drive, flying memory).